rename header file to better conform to standards
