+++ examples/c_usage_encoding_example.c
@@ -1,4 +1,4 @@
-#include "libbech32.h"
+#include "bech32.h"
 #include <string.h>
 #include <stdio.h>
