+++ examples/cpp_other_examples.cpp
@@ -1,4 +1,4 @@
-#include "libbech32.h"
+#include "bech32.h"
 #include <iostream>
 
 // make sure we can check these examples even when building a release version
